Update Admin
-Dashboard
-Manage User(Makedmin, delete user)
-Manage Post
-Manage Comment
-Filter
-Search

+++ api/controllers/post.controller.js
@@ -1,31 +1,24 @@
-import Post from "../models/post.model.js";
-import { errorHandler } from "../utils/error.js";
+import Post from '../models/post.model.js';
+import { errorHandler } from '../utils/error.js';
 
 export const create = async (req, res, next) => {
-  const { title, content, topicID } = req.body;
-
-  if (!title || !content) {
-    return next(errorHandler(400, "Please provide all required fields"));
+  if (!req.user.isAdmin) {
+    return next(errorHandler(403, 'You are not allowed to create a post'));
   }
-
-  const slug = title
-    .split(" ")
-    .join("-")
+  if (!req.body.title || !req.body.content) {
+    return next(errorHandler(400, 'Please provide all required fields'));
+  }
+  const slug = req.body.title
+    .split(' ')
+    .join('-')
     .toLowerCase()
-    .replace(/[^a-zA-Z0-9-]/g, "");
-
+    .replace(/[^a-zA-Z0-9-]/g, '');
+  const newPost = new Post({
+    ...req.body,
+    slug,
+    userId: req.user.id,
+  });
   try {
-    const existingPost = await Post.findOne({ slug });
-    if (existingPost) {
-      return next(errorHandler(400, "A post with this slug already exists. Please choose a different title."));
-    }
-
-    const newPost = new Post({
-      ...req.body,
-      slug,
-      userId: req.user.id,
-    });
-
     const savedPost = await newPost.save();
     res.status(201).json(savedPost);
   } catch (error) {
@@ -37,7 +30,7 @@ export const getposts = async (req, res, next) => {
   try {
     const startIndex = parseInt(req.query.startIndex) || 0;
     const limit = parseInt(req.query.limit) || 9;
-    const sortDirection = req.query.order === "asc" ? 1 : -1;
+    const sortDirection = req.query.order === 'asc' ? 1 : -1;
     const posts = await Post.find({
       ...(req.query.userId && { userId: req.query.userId }),
       ...(req.query.category && { category: req.query.category }),
@@ -45,8 +38,8 @@ export const getposts = async (req, res, next) => {
       ...(req.query.postId && { _id: req.query.postId }),
       ...(req.query.searchTerm && {
         $or: [
-          { title: { $regex: req.query.searchTerm, $options: "i" } },
-          { content: { $regex: req.query.searchTerm, $options: "i" } },
+          { title: { $regex: req.query.searchTerm, $options: 'i' } },
+          { content: { $regex: req.query.searchTerm, $options: 'i' } },
         ],
       }),
     })
@@ -58,7 +51,11 @@ export const getposts = async (req, res, next) => {
 
     const now = new Date();
 
-    const oneMonthAgo = new Date(now.getFullYear(), now.getMonth() - 1, now.getDate());
+    const oneMonthAgo = new Date(
+      now.getFullYear(),
+      now.getMonth() - 1,
+      now.getDate()
+    );
 
     const lastMonthPosts = await Post.countDocuments({
       createdAt: { $gte: oneMonthAgo },
@@ -75,20 +72,20 @@ export const getposts = async (req, res, next) => {
 };
 
 export const deletepost = async (req, res, next) => {
-  if (req.user.id !== req.params.userId) {
-    return next(errorHandler(403, "You are not allowed to delete this post"));
+  if (!req.user.isAdmin || req.user.id !== req.params.userId) {
+    return next(errorHandler(403, 'You are not allowed to delete this post'));
   }
   try {
     await Post.findByIdAndDelete(req.params.postId);
-    res.status(200).json("The post has been deleted");
+    res.status(200).json('The post has been deleted');
   } catch (error) {
     next(error);
   }
 };
 
 export const updatepost = async (req, res, next) => {
-  if (req.user.id !== req.params.userId) {
-    return next(errorHandler(403, "You are not allowed to update this post"));
+  if (!req.user.isAdmin || req.user.id !== req.params.userId) {
+    return next(errorHandler(403, 'You are not allowed to update this post'));
   }
   try {
     const updatedPost = await Post.findByIdAndUpdate(
